use robin hood hashing for syncmers directly

+++ src/main/taxor_build_configuration.hpp
@@ -10,9 +10,9 @@ struct configuration
     std::string input_file_name{}; // provided by user
     std::string input_sequence_folder{};
     std::string output_file_name{}; // provided by user
-    uint8_t threads{1u};
-    uint8_t kmer_size{20u};
-    uint8_t syncmer_size{10u};
+    int threads{1u};
+    int kmer_size{20u};
+    int syncmer_size{10u};
     bool output_verbose_statistics{false};
     bool debug{false};
     bool use_syncmer{false};
